Desarrollo Negocio.

+++ Dominio/Persona.cs
@@ -11,10 +11,10 @@ namespace Dominio
         public Int64 IdPersona { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Dni { get; set; }
-        public int Cuit { get; set; }
+        public Int64 Dni { get; set; }
+        public Int64 Cuit { get; set; }
         public bool TipoPersona { get; set; }
-        public int Telefono { get; set; }
+        public Int64 Telefono { get; set; }
         public string Email { get; set; }
         public string Direccion { get; set; }
     }
